Realize that pushing turbo a when jumping gives you really bad results
So seriously. If your problems seem to be due to spurious inputs, and
you don't realize that your problem is that you're using the turbo A
button on your emulator instead of the A button, how dumb are you going
to feel when you figure it out?

That's me....

It also appears that the jump actually is overflowing the velocity because
every time I tested it it would immediately cancel due to turbo A...

facepalm