Mess around with screen transitions a bit
I do have one idea about how these could really be improved.

Basically, we would copy the new map into the other nametable, and
then switch nametables immediately. However, we can only copy data
either while the background is off, or during vblank. So we would
have to copy the data into a buffer, then copy it to the PPU during
vblank, and we might have to use a few frames to update the nametable.

However, this would let us keep drawing player sprites while we were
performing the update.

In fact, we could trigger the screen transition a little early,
allowing the player to have an almost continuous experience.

We could even try to start loading the map a little early, i.e. while
the player is quite a distance aways from the edge of the screen.
However this probably wouldn't work because in any quadrant of the
screen there can be up to two maps that might get loaded next.

I think the best idea would be to trigger map loading while the player
is starting to move into the edge of the screen, and then maybe cancel
it if the player turns around really fast...? The question is whether
all that loading and buffering will make the player not move correctly,
which would actually be more disruptive than just doing the blank frame
as we're doing now.